split out SetFilters
